test: factor test strcuture in playground
