+++ packages/untp-playground/tsconfig.json
@@ -34,7 +34,8 @@
     "**/*.tsx", 
     ".next/types/**/*.ts", 
     "__tests__/**/*.ts", 
-    "__tests__/**/*.tsx"
+    "__tests__/**/*.tsx",
+    "jest.config.js"
   ],
   "exclude": [
     "node_modules",
